In some cases. `buildTime` used in all JARs' manifest caused all subprojects to be rebuilt and retested.

## Results (for Groovy 3):
### Before
```
$ gw clean check
...
BUILD SUCCESSFUL in 1m 20s
70 actionable tasks: 68 executed, 2 up-to-date
```
No changes:
```
$ gw check
...
BUILD SUCCESSFUL in 1m 15s
54 actionable tasks: 32 executed, 22 up-to-date
```
Change in spock-spec (all tests in that module are re-executed):
```
$ gw check
...
BUILD SUCCESSFUL in 1m 19s
54 actionable tasks: 32 executed, 22 up-to-date
```

### After
```
$ gw clean check
...
BUILD SUCCESSFUL in 1m 20s
70 actionable tasks: 68 executed, 2 up-to-date
```
No changes:
```
$ gw check
...
BUILD SUCCESSFUL in 1s
54 actionable tasks: 54 up-to-date
```
Change in spock-spec (all tests in that module are re-executed):
```
$ gw check
...
BUILD SUCCESSFUL in 39s
54 actionable tasks: 2 executed, 52 up-to-date
```

We may change the way how snapshot release build is detected. Currently it's only Travis, but I disabled this optimization for all CI (that set `CI` env var :-) ).

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/spockframework/spock/1110)
<!-- Reviewable:end -->